Update script.js
new object combinedMovementsDate in displayMovements. It fixes the bug when  sorting the movements array. It was not taking in consideration that the indexes of the movements array were different from the movementsDate array so the latter would always be display in the same order, ignoring the new index order of the movements array